【fix】基础修改

+++ Unity/Assets/Scripts/Model/Generate/ClientServer/Config/RewardConfig.cs
@@ -18,7 +18,7 @@ namespace ET
         {
             Id = _buf.ReadInt();
             Name = _buf.ReadString();
-            {int n0 = System.Math.Min(_buf.ReadSize(), _buf.Size);Reward = new System.Collections.Generic.Dictionary<int, int>(n0 * 3 / 2);for(var i0 = 0 ; i0 < n0 ; i0++) { int _k0;  _k0 = _buf.ReadInt(); int _v0;  _v0 = _buf.ReadInt();     Reward.Add(_k0, _v0);}}
+            {int n0 = System.Math.Min(_buf.ReadSize(), _buf.Size);Reward = new System.Collections.Generic.Dictionary<int, long>(n0 * 3 / 2);for(var i0 = 0 ; i0 < n0 ; i0++) { int _k0;  _k0 = _buf.ReadInt(); long _v0;  _v0 = _buf.ReadLong();     Reward.Add(_k0, _v0);}}
 
             PostInit();
         }
@@ -41,7 +41,7 @@ namespace ET
         /// <summary>
         /// 奖励包内容
         /// </summary>
-        public readonly System.Collections.Generic.Dictionary<int, int> Reward;
+        public readonly System.Collections.Generic.Dictionary<int, long> Reward;
 
         public const int __ID__ = 1443672945;
